Dev/add config

+++ src/Widget/WidgetLayout/Header/index.jsx
@@ -1,4 +1,5 @@
-import { Bars3BottomRightIcon } from "@heroicons/react/24/outline";
+import { Bars3BottomRightIcon } from "@heroicons/react/24/outline"; // Add CheckIcon or any other icon
+import { FaSave, FaSync } from "react-icons/fa"; // Import your icon
 import { useContext, useEffect, useState } from "react";
 import AppContext from "../../AppContext";
 import { motion } from "framer-motion";
@@ -6,32 +7,22 @@ import { useDetectClickOutside } from "../../../hooks/useDetectClickOutside";
 import { useDispatch, useSelector } from "react-redux";
 import {
   roleMap,
-  setRemindTime,
   setRole,
-  setToggleWidget,
   setToken,
 } from "../../widgetSlice";
 import {
+  fetchChatHistory,
   removeAllMessages,
   resetBot,
-  resetMessageState,
+  setRemindApi,
   setUserTypingPlaceholder,
   toggleBotTyping,
   toggleUserTyping,
 } from "../Messages/messageSlice";
 import { Icon } from "./Icons";
-
-const dropdownMenu = [
-  {
-    title: "Restart",
-  },
-  {
-    title: "Clear Chat",
-  },
-  {
-    title: "Close",
-  },
-];
+import { IconButton } from "./IconButton";
+import Swal from "sweetalert2";
+import { toast } from "react-toastify";
 
 export const mapRole = ["Default", "Professor", "Assistant", "Friend"];
 
@@ -39,7 +30,9 @@ const models = ["Chat GPT 3.5", "Chat GPT 4"];
 
 export const Header = () => {
   const dispatch = useDispatch();
-  let { role, remindTime, token } = useSelector((state) => state.widgetState);
+  let { role, remindTime, remind, token } = useSelector(
+    (state) => state.widgetState
+  );
   const appContext = useContext(AppContext);
 
   const {
@@ -53,8 +46,10 @@ export const Header = () => {
 
   const { textColor, backgroundColor, enableBotAvatarBorder } = chatHeaderCss;
   const [showDropdown, setShowDropdown] = useState(false);
-  const [remind, setRemind] = useState(false);
+
   const [selectedModel, setSelectedModel] = useState("Chat GPT 3.5");
+  const [remindState, setRemindState] = useState(remind);
+  const [remindTimeState, setRemindTimeState] = useState(remindTime);
   const dropdownRef = useDetectClickOutside({
     setShowModal: setShowDropdown,
   });
@@ -63,13 +58,14 @@ export const Header = () => {
     const roleKey = event.target.value;
     dispatch(setRole(roleKey)); // Store the key (1, 2, 3) in context
   };
+
   // Retrieve the token from localStorage on component mount
   useEffect(() => {
     const storedToken = localStorage.getItem("token");
     if (storedToken) {
       dispatch(setToken(storedToken));
     }
-  }, [token]);
+  }, []);
 
   const handleTokenChange = (event) => {
     const tokenValue = event.target.value;
@@ -78,37 +74,71 @@ export const Header = () => {
   };
 
   const handleRemindToggle = () => {
-    setRemind(!remind);
+    setRemindState((prev) => !prev);
   };
 
   const handleRemindTimeChange = (event) => {
-    dispatch(setRemindTime(event.target.value)); // Update the context
+    setRemindTimeState(event.target.value); // Update the context
   };
 
   const handleModelChange = (event) => {
     setSelectedModel(event.target.value);
   };
 
-  const handleCloseButton = () => {
-    dispatch(setToggleWidget(false));
-    setShowDropdown(!showDropdown);
-  };
-
   const handleClearChatButton = () => {
-    dispatch(removeAllMessages());
-    dispatch(toggleBotTyping(false));
-    dispatch(toggleUserTyping(true));
-    dispatch(setUserTypingPlaceholder("Type your message..."));
-    setShowDropdown(!showDropdown);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        setShowDropdown(!showDropdown);
+        dispatch(removeAllMessages());
+        dispatch(toggleBotTyping(false));
+        dispatch(toggleUserTyping(true));
+        dispatch(setUserTypingPlaceholder("Type your message..."));
+        dispatch(
+          resetBot({
+            rasaServerUrl: `${rasaServerUrl}/chat?chatid=${userId}`,
+            token: token,
+          })
+        );
+      }
+    });
   };
 
-  const handleRestartButton = () => {
-    dispatch(resetMessageState());
-    setShowDropdown(!showDropdown);
-    dispatch(
-      resetBot({
-        rasaServerUrl: `${rasaServerUrl}?chatid=${userId}`,
+  useEffect(() => {
+    setRemindState(remind);
+    setRemindTimeState(remindTime);
+  }, [remind, remindTime]);
+
+  const handleSyncChatHistory = async () => {
+    const res = await dispatch(
+      fetchChatHistory({
+        rasaServerUrl: `${rasaServerUrl}/chat?chatid=${userId}`,
         token: token,
+      })
+    );
+    if (res.payload.error) {
+      toast.error("Sync Chat failed!")
+    }
+    if (res.payload.message) {
+      toast.success("Sync Chat successfully!")
+    }
+  };
+
+  const handleSaveRemind = async () => {
+    await dispatch(
+      setRemindApi({
+        rasaServerUrl: `${rasaServerUrl}/settime`,
+        token: token,
+        userId: userId,
+        status: remindState,
+        remindTime: remindTimeState,
       })
     );
   };
@@ -154,6 +184,24 @@ export const Header = () => {
             }}
           >
             <li className="p-2">
+              <label htmlFor="model" className="mr-2">
+                Model:
+              </label>
+              <select
+                id="model"
+                value={selectedModel}
+                onChange={handleModelChange}
+                className="rounded-lg border p-1"
+                style={{ color: textColor, borderColor: textColor }}
+              >
+                {models.map((model, idx) => (
+                  <option key={idx} value={model}>
+                    {model}
+                  </option>
+                ))}
+              </select>
+            </li>
+            <li className="p-2">
               <label htmlFor="role" className="mr-2">
                 Role:
               </label>
@@ -175,14 +223,27 @@ export const Header = () => {
               <label htmlFor="token" className="mr-2">
                 Token:
               </label>
-              <input
-                id="token"
-                type="text"
-                value={token}
-                onChange={handleTokenChange}
-                className="rounded-lg border p-1"
-                style={{ color: textColor, borderColor: textColor }}
-              />
+              <div className="flex items-center">
+                <input
+                  id="token"
+                  type="text"
+                  value={token}
+                  onChange={handleTokenChange}
+                  className="rounded-lg border p-1"
+                  style={{ color: textColor, borderColor: textColor }}
+                />
+                <div className="group relative inline-block">
+                  <IconButton
+                    icon={FaSync}
+                    onClick={handleSyncChatHistory}
+                    disabled={!token}
+                    tooltip={"Sync Chat History"}
+                  />
+                  <span className="absolute bottom-full left-1/2 z-50 mb-2 w-max -translate-x-1/2 transform rounded-md bg-gray-700 p-1 text-xs text-white opacity-0 transition-opacity duration-300 group-hover:opacity-100">
+                    {"Sync Chat history"}
+                  </span>
+                </div>
+              </div>
             </li>
             <li className="p-2">
               <div className="flex items-center">
@@ -192,63 +253,44 @@ export const Header = () => {
                 <input
                   type="checkbox"
                   id="remind"
-                  checked={remind}
+                  checked={remindState}
                   onChange={handleRemindToggle}
                 />
-                {remind && (
+                {remindState && userId === "2" && (
                   <input
                     type="time"
-                    value={remindTime}
+                    value={remindTimeState}
                     onChange={handleRemindTimeChange}
                     className="ml-2 rounded-lg border p-1"
                     style={{ color: textColor, borderColor: textColor }}
                   />
                 )}
+                <div className="ml-2">
+                  <IconButton
+                    icon={FaSave} // You can change this to your actual save icon
+                    onClick={handleSaveRemind}
+                    disabled={false} // Adjust this based on your logic
+                    tooltip={"Save"}
+                  />
+                </div>
               </div>
             </li>
+
             <li className="p-2">
-              <label htmlFor="model" className="mr-2">
-                Model:
-              </label>
-              <select
-                id="model"
-                value={selectedModel}
-                onChange={handleModelChange}
-                className="rounded-lg border p-1"
-                style={{ color: textColor, borderColor: textColor }}
+              <div
+                className="flex cursor-pointer hover:opacity-70"
+                onClick={() => handleClearChatButton()}
               >
-                {models.map((model, idx) => (
-                  <option key={idx} value={model}>
-                    {model}
-                  </option>
-                ))}
-              </select>
-            </li>
-            {dropdownMenu.map((item, idx) => {
-              const { title } = item;
-              return (
-                <div
-                  key={idx}
-                  className="flex hover:opacity-70"
-                  onClick={() => {
-                    if (title === "Close") {
-                      handleCloseButton();
-                    } else if (title === "Clear Chat") {
-                      handleClearChatButton();
-                    } else {
-                      handleRestartButton();
-                    }
-                  }}
-                >
-                  <div className="flex items-center justify-center pl-2">
-                    <Icon name={title} />
-                  </div>
-                  <div>
-                    <span className="block py-2 px-2">{title}</span>
-                  </div>
+                <div className="flex items-center justify-center pl-2">
+                  <Icon name={"Clear Chat"} />
                 </div>
-              );
-            })}
+                <div>
+                  <span className="block py-2 px-2">{"Clear Chat"}</span>
+                </div>
+              </div>
+
+              {/* Add Save button below Clear Chat */}
+            </li>
           </ul>
         </div>
       )}
